Update package.json and remove duplicate dependencies
