+++ package.json
@@ -18,8 +18,6 @@
     "framer-motion": "^11.11.8",
     "jspdf": "^2.5.2",
     "jspdf-autotable": "^3.8.3",
-    "jspdf": "^2.5.2",
-    "jspdf-autotable": "^3.8.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.27.0",
